+++ gemsiii/accurate_scan/fixpoint.h
@@ -12,7 +12,7 @@
 
 typedef int fixpoint;
 #include <stdint.h>
-typedef struct {uint64_t hi, lo, neg;} dblfixpoint;
+typedef struct { int64_t hi, lo, neg;} dblfixpoint;
 
 extern int fp_error;
 extern  fixpoint fp_max();
@@ -26,7 +26,7 @@ extern  void fp_print();
 extern  fixpoint fp_fix();
 extern double fp_double(fixpoint x);
 
-extern int fp_dblnegative();
+extern uint64_t fp_dblnegative();
 extern dblfixpoint fp_dblnegate();
 extern dblfixpoint fp_dblmultiply();
 extern int fp_dbllessthan();
